Asset Class to hold an asset

Asset Class to hold an asset

Asset Class to hold an asset

Active bindings

uid

uid of the asset.

name

name of the asset.

asset_url

asset_url of the asset. takes the form of assets/uid/ Contains most of the information about an asset such as meta data, permissions, xls form content and links to other related endpoints.

data_url

data_url of the asset. This can be used to get data related to the asset. Most useful for the asset type survey where this URL gives access to the submissions to the survey.

owner_username

username of the owner of the asset.

type

type of the asset. Type of the asset, e.g. survey, question, block or template.

Methods

Public methods

Asset$new()

Asset$get_submissions()

Asset$to_list()

Asset$clone()

Method new()

Usage

Asset$new(asset_list, kobo)

Arguments

asset_list

list. a list with the asset data, such as obtained through a call to assets/id

kobo

Kobo instance. Instance of class Kobo used internally to make requests to the API. get_submissions

Method get_submissions()

get submissions to a survey/form

Usage

Asset$get_submissions()

Returns

tibble. submissions as a tibble. if no submissions were made yet, the tibble will have no columns. to_list

Method to_list()

Usage

Asset$to_list()

Returns

list representation of asset

Method clone()

The objects of this class are cloneable with this method.

Usage

Asset$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.