Kobo Class

Interface object for the Kobo API that can handle KoboClient instances (sessions) for both API versions. The Class exposes both generic and specific methods for HTTP requests / interactions with the various endpoints.

Methods

Public methods

Kobo$new()

Kobo$get()

Kobo$get_assets()

Kobo$clone()

Method new()

Initialization method for class "Kobo".

Usage

Kobo$new(
  base_url_v2,
  base_url_v1 = NULL,
  kobo_token = Sys.getenv("KBTBR_TOKEN")
)

Arguments

base_url_v2

character. The base URL of the API version 2 (known as /api/v2). For example: https://kobo.correlaid.org.

base_url_v1

character. The base URL of the API of your KoBoCAT API (also known as /api/v1). Defaults to NULL. For example: https://kc.correlaid.org.

kobo_token

character. The API token. Defaults to requesting the systen environment KBTBR_TOKEN.

Method get()

Wrapper for the GET method of internal session objects.

Usage

Kobo$get(path, query = list(format = "json"), version = "v2")

Arguments

path

character. Path component of the endpoint.

query

list. A named list which is parsed to the query component. The order is not hierarchical.

version

character. Indicates on which API version the request should be executed (available: v1, v2). Defaults to v2.

Method get_assets()

Example method to send a GET request to the assets endpoint (due to default to v2, no further specification is needed).

Usage

Kobo$get_assets()

Method clone()

The objects of this class are cloneable with this method.

Usage

Kobo$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.